PoissonModel at least not crashing.
git-svn-id: https://svn.code.sf.net/p/jaer/code/trunk/host/java@3695 b7f4320f-462c-0410-a916-d9f35bb82d52

+++ src/ch/unizh/ini/jaer/projects/apsdvsfusion/firingmodel/IntegerDecayModel.java
@@ -91,7 +91,7 @@ public class IntegerDecayModel {
 	
 	public void decay(int timePassed) {
 		final int reductions = timePassed / tau;
-		value *= multValues[(int)(((timePassed - (reductions * tau))) << multValueCounterShift) / tau];
+		value *= multValues[(int)((((long)(timePassed - (reductions * tau))) << multValueCounterShift) / tau)];
 		value >>= (reductions+multValueDivisorShift);
 	}
